requirements.txt
pillow==11.3.0
pure-python-adb==0.3.0.dev0